Added `dynaconf list` and `dynaconf write` subcommands

+++ tests/test_cli.py
@@ -3,6 +3,7 @@ import pytest
 from pathlib import Path
 from click.testing import CliRunner
 from dynaconf.cli import main, EXTS
+from dotenv import cli as dotenv_cli
 
 
 runner = CliRunner()
@@ -24,56 +25,74 @@ def test_banner():
 
 @pytest.mark.parametrize("fileformat", EXTS)
 def test_init(fileformat):
+    if fileformat == 'env':
+        path = '.env'
+        secs_path = None
+    else:
+        path = 'settings.{}'.format(fileformat)
+        secs_path = '.secrets.{}'.format(fileformat)
+
     run(
         [
             'init',
-            '-no-wg',
+            '--no-wg',
             '--format={}'.format(fileformat),
             '-v name=bruno',
             '-y'
         ]
     )
 
-    sets = Path('settings.{}'.format(fileformat))
-    secs = Path('.secrets.{}'.format(fileformat))
-    # gign = Path('.gitignore')
-
+    sets = Path(path)
     assert sets.exists() is True
-    assert secs.exists() is True
-    # assert gign.exists() is True
-
     assert 'bruno' in open(sets).read()
-    assert 'secret for' in open(secs).read()
-    # assert ".secrets.*" in open(gign).read()
 
-    os.remove(sets)
-    os.remove(secs)
+    if fileformat != 'env':
+        os.remove(sets)
+    else:
+        dotenv_cli.unset_key(path, 'NAME')
+
+    if secs_path:
+        secs = Path('.secrets.{}'.format(fileformat))
+        assert secs.exists() is True
+        assert 'secret for' in open(secs).read()
+        os.remove(secs)
+
+    # gign = Path('.gitignore')
+    # assert gign.exists() is True
+    # assert ".secrets.*" in open(gign).read()
 
 
 @pytest.mark.parametrize("fileformat", EXTS)
 def test_init_with_path(fileformat):
     # run twice to force load of existing files
+    if fileformat == 'env':
+        path = '/tmp/.env'
+        secs_path = None
+    else:
+        path = '/tmp/settings.{}'.format(fileformat)
+        secs_path = '/tmp/.secrets.{}'.format(fileformat)
+
     for _ in (1, 2):
         run(
             [
-                'init', '-no-wg',
+                'init', '--no-wg',
                 '--format={}'.format(fileformat),
-                '--path=/tmp/settings.{}'.format(fileformat),
+                '--path={}'.format(path),
                 '-y'
             ]
         )
 
-    sets = Path('/tmp/settings.{}'.format(fileformat))
-    secs = Path('/tmp/.secrets.{}'.format(fileformat))
-    # gign = Path('.gitignore')
-
+    sets = Path(path)
     assert sets.exists() is True
-    assert secs.exists() is True
-    # assert gign.exists() is True
-
     assert 'value for default' in open(sets).read()
-    assert 'secret for' in open(secs).read()
-    # assert ".secrets.*" in open(gign).read()
-
     os.remove(sets)
-    os.remove(secs)
+
+    if secs_path:
+        secs = Path('/tmp/.secrets.{}'.format(fileformat))
+        assert secs.exists() is True
+        assert 'secret for' in open(secs).read()
+        os.remove(secs)
+
+    # gign = Path('.gitignore')
+    # assert gign.exists() is True
+    # assert ".secrets.*" in open(gign).read()
